refactor(credit): remove credit repository and update imports to creditTransaction
