+++ web-app/src/lib/db/repositories/index.ts
@@ -1,7 +1,7 @@
 export * from "./agent";
 export * from "./agentList";
-export * from "./credit";
 export * from "./creditCost";
+export * from "./creditTransaction";
 export * from "./fiatTransaction";
 export * from "./invitation";
 export * from "./job";
